requirements.txt
# FastAPI core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP client (for external APIs)
httpx==0.26.0

# Data processing
pandas==2.2.1
numpy<2

# Finance data
akshare==1.18.55

# Password hashing
bcrypt==4.0.1

# Production server
gunicorn==21.2.5